chore(experiments): rename CharModal legacy component file.
