+++ frontend/src/scenes/experiments/legacy/metricsView/LegacyChartModal.tsx
